pyproject.toml
[tool.poetry]
name = "kvecdb"
version = "0.1.0"
description = ""
authors = ["Arsalan Azmi <mrlethalazo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.24.3"
pandas = "^2.0.1"
sentence-transformers = "^2.2.2"
torch = "^2.0.1"
scikit-learn = "^1.2.2"
ipywidgets = "^8.0.6"
qPython3 = "^1.0.1"

[tool.poetry.dev-dependencies]
ipykernel = "^6.23.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
